Move testbed command to benchmarks
